requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.2.5
autopep8==1.4.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
gast==0.2.2
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
isort==4.3.20
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.3.1
pypiwin32==223
pywin32==227
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
scipy==1.4.1
six==1.12.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-addons==0.8.2
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.42.1
typed-ast==1.4.0
typeguard==2.7.1
urllib3==1.25.7
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.1